Move configs to just one file
