--- src/lightning_hydra_zen_template/configs/hydra_conf/default.py
@@ -1,24 +1,0 @@
-import os
-
-from hydra.conf import HydraConf, RunDir, SweepDir
-from hydra_zen import store
-
-log_dir = "${paths.log_dir}"
-output_dir = "${paths.output_dir}"
-yyyymmdd = "${now:%Y-%m-%d}"
-hhmmss = "${now:%H-%M-%S}"
-task_name = "${task_name}"
-
-run_dir = os.path.join(log_dir, task_name, "runs", yyyymmdd, hhmmss)
-sweep_dir = os.path.join(log_dir, task_name, "multiruns", yyyymmdd, hhmmss)
-job_file = os.path.join(output_dir, f"{task_name}.log")
-
-HydraCfg = HydraConf(
-    run=RunDir(run_dir),
-    sweep=SweepDir(dir=sweep_dir, subdir="${hydra:job.num}"),
-    # callbacks={"print_config": builds(PrintConfigCallback)},
-    # Fix from PR https://github.com/facebookresearch/hydra/pull/2242, while there isn't a new release
-    job_logging={"handlers": {"file": {"filename": job_file}}},
-)
-
-store(HydraCfg)
