+++ src/lightning_hydra_zen_template/utils/print_config.py
@@ -1,16 +1,19 @@
 import copy
-import logging
 import os
 
 import rich
 import rich.syntax
 import rich.tree
+from lightning_utilities.core.rank_zero import rank_zero_only
 from omegaconf import DictConfig, OmegaConf, flag_override
 
-log = logging.getLogger(__name__)
+from lightning_hydra_zen_template.utils.ranked_logger import RankedLogger
+
+log = RankedLogger(__name__, rank_zero_only=True)
 
 
-def save_and_print_cfg(cfg: DictConfig) -> None:
+@rank_zero_only
+def print_config(cfg: DictConfig) -> None:
     file_path = os.path.join(cfg.paths.output_dir, "resolved_config.yaml")
     log.info(f"Saving config to {file_path}")
     cfg = OmegaConf.create(OmegaConf.to_container(cfg, resolve=True))
@@ -26,7 +29,21 @@ def save_and_print_cfg(cfg: DictConfig) -> None:
 
     tree = rich.tree.Tree("CONFIG")
 
-    for key, value in cfg.items():
+    print_order = ["data", "model", "trainer"]
+    queue = []
+
+    for field in print_order:
+        if field in cfg:
+            queue.append(field)
+        else:
+            log.warning(f"Field '{field}' not found in config. Skipping '{field}' config printing...")
+
+    for field in cfg:
+        if field not in queue:
+            queue.append(field)
+
+    for key in queue:
+        value = cfg[key]
         is_dict = isinstance(value, DictConfig)
         convert_fn = OmegaConf.to_yaml if is_dict else str
         branch_content = convert_fn(value)
